rlg-backend/requirements.txt
# FastAPI Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# Local Embeddings (offline)
sentence-transformers>=2.2.0
torch>=2.0.0

# Local Vector Search
faiss-cpu>=1.7.4

# Local LLM Integration
ollama>=0.1.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.2
beautifulsoup4>=4.12.0

# OCR (offline - optional)
pytesseract>=0.3.10
pdf2image>=1.17.0

# Text Processing
rank-bm25>=0.2.2

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
